+++ R/check_lintr.R
@@ -1,17 +1,5 @@
-#' @title The lintr check
-#'
-#' @description Runs lint from the \code{lintr} package. Also scans for all
-#' \code{.Rmd} and \code{.R} files in other directories.
-#' @importFrom lintr lint_package lint_dir lint
-#' @export
-check_lintr = function() {
-  cli::cli_h3("Checking lint...check_lintr()")
+lint_files = function() {
   lint_errors = FALSE
-
-  if (!file.exists(".lintr")) {
-    cli::cli_alert_info("No .lintr file found")
-  }
-
   lints = lintr::lint_package()
   if (length(lints) > 0) {
     lapply(lints, print)
@@ -27,6 +15,27 @@ check_lintr = function() {
       lint_errors = TRUE
     }
   }
+
+  return(lint_errors)
+}
+
+#' @title The lintr check
+#'
+#' @description Runs lint from the \code{lintr} package. Also scans for all
+#' \code{.Rmd} and \code{.R} files in other directories.
+#' @importFrom lintr lint_package lint_dir lint
+#' @export
+check_lintr = function() {
+  cli::cli_h3("Checking lint...check_lintr()")
+
+  if (!file.exists(".lintr")) {
+    cli::cli_alert_info("No .lintr file found")
+  }
+
+  lint_errors = lint_files()
+  # Ensure files end with newlines
+  rfiles = list.files("R", full.names = TRUE)
+  lapply(rfiles, readLines) #Uses inteRgrate readlines
 
   if (isTRUE(lint_errors)) {
     msg_error("Please fix linting errors", stop = TRUE)