+++ R/read_lines.R
@@ -1,4 +1,4 @@
-# WIP: https://community.rstudio.com/t/incomplete-final-line-found-whats-the-big-deal/56536
+# https://community.rstudio.com/t/incomplete-final-line-found-whats-the-big-deal/56536
 # Also cleans up noise from readLines
 # nolint start
 readLines = function(con = stdin(), n = -1L, ok = TRUE, warn = TRUE,
@@ -8,9 +8,8 @@ readLines = function(con = stdin(), n = -1L, ok = TRUE, warn = TRUE,
                error = function(e) e,
                warning = function(w) w)
   if ("warning" %in% class(out)) {
-    msg = paste0("Files should end with an empty line. Check \n\n", normalizePath(con))
-    msg_error(msg)
-    stop(call. = FALSE)
+    msg = paste0("Files should end with an empty line. Check \n", normalizePath(con))
+    msg_error(msg, stop = TRUE)
   }
   out
 }