+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(inteRgrate)
+
+test_check("inteRgrate")